Remove Chris.AI.EQS.FieldView
