+++ AI/Runtime/EQS/Component/FieldViewPrimeQueryComponent.cs
@@ -4,10 +4,41 @@ using UnityEngine;
 using UnityEngine.Pool;
 namespace Chris.AI.EQS
 {
+    public interface IFieldViewQueryComponent
+    {
+        /// <summary>
+        /// Request a new query from target's field view
+        /// </summary>
+        /// <returns></returns>
+        bool RequestFieldViewQuery();
+
+        /// <summary>
+        /// Detect whether it can see the target
+        /// </summary>
+        /// <param name="target"></param>
+        /// <param name="fromPosition"></param>
+        /// <param name="fromRotation"></param>
+        /// <param name="filterTags"></param>
+        /// <returns></returns>
+        bool Detect(Vector3 target, Vector3 fromPosition, Quaternion fromRotation, string[] filterTags = null);
+        
+        /// <summary>
+        /// Query actors overlap in field of view from cache
+        /// </summary>
+        /// <param name="actors"></param>
+        void CollectViewActors(List<Actor> actors);
+        
+        /// <summary>
+        /// Query actors overlap in field of view from cache
+        /// </summary>
+        /// <param name="actors"></param>
+        void CollectViewActors<T>(List<T> actors) where T : Actor;
+    }
+    
     /// <summary>
     /// Field view prime query data provider associated with an Actor as component
     /// </summary>
-    public class FieldViewPrimeQueryComponent : FieldViewQueryComponentBase
+    public class FieldViewPrimeQueryComponent : ActorComponent, IFieldViewQueryComponent
     {
         [Header("Data")]
         public FieldViewPrime fieldView = new()
@@ -34,7 +65,7 @@ namespace Chris.AI.EQS
             }
         }
         
-        public override bool RequestFieldViewQuery()
+        public bool RequestFieldViewQuery()
         {
             if (_system == null)
             {
@@ -49,12 +80,12 @@ namespace Chris.AI.EQS
             return true;
         }
         
-        public override void CollectViewActors(List<Actor> actors)
+        public void CollectViewActors(List<Actor> actors)
         {
             _system.GetActorsInFieldView(GetActor().GetActorHandle(), actors);
         }
         
-        public override void CollectViewActors<T>(List<T> actors)
+        public void CollectViewActors<T>(List<T> actors) where T : Actor
         {
             var list = ListPool<Actor>.Get();
             CollectViewActors(list);
@@ -65,7 +96,7 @@ namespace Chris.AI.EQS
             ListPool<Actor>.Release(list);
         }
         
-        public override bool Detect(Vector3 target, Vector3 fromPosition, Quaternion fromRotation, string[] filterTags = null)
+        public bool Detect(Vector3 target, Vector3 fromPosition, Quaternion fromRotation, string[] filterTags = null)
         {
             return fieldView.Detect(target, fromPosition, fromRotation, queryLayerMask, filterTags);
         }
