--- AI/Runtime/EQS/Component/FieldViewQueryComponent.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c63c5e65e49adcc4d90c9806b47a7420
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
