--- AI/Runtime/EQS/Model/FieldView.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b0c3fc61497d0874da755bf349c57c29
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
